fix type

+++ app/src/main/java/com/nakano/forecastmvvm/data/db/entity/CurrentWeatherEntry.kt
@@ -9,12 +9,12 @@ const val CURRENT_WEATHER_ID = 0
 
 @Entity(tableName = "current_weather")
 data class CurrentWeatherEntry(
-    val feelslike: Int,
+    val feelslike: Double,
     @SerializedName("is_day")
     val isDay: String,
-    val precip: Int,
-    val temperature: Int,
-    val visibility: Int,
+    val precip: Double,
+    val temperature: Double,
+    val visibility: Double,
     @SerializedName("weather_code")
     val weatherCode: Int,
 //    @SerializedName("weather_descriptions")
@@ -24,7 +24,7 @@ data class CurrentWeatherEntry(
     @SerializedName("wind_dir")
     val windDir: String,
     @SerializedName("wind_speed")
-    val windSpeed: Int
+    val windSpeed: Double
 ){
     @PrimaryKey(autoGenerate = false)
     var id: Int = CURRENT_WEATHER_ID
